+++ lib/trailblazer/activity/fast_track.rb
@@ -106,11 +106,6 @@ module Trailblazer
           ctx
         end
 
-        def initial_sequence(sequence:, fail_fast_end: Activity::End.new(semantic: :fail_fast), pass_fast_end: Activity::End.new(semantic: :pass_fast), **)
-          sequence = Path::DSL.append_terminus(sequence, fail_fast_end, magnetic_to: :fail_fast, id: "End.fail_fast", normalizers: Normalizers)
-          sequence = Path::DSL.append_terminus(sequence, pass_fast_end, magnetic_to: :pass_fast, id: "End.pass_fast", normalizers: Normalizers)
-        end
-
         # Normalizer pipelines taking care of processing your DSL options.
         Normalizers = Linear::Normalizer::Normalizers.new(
           step: FastTrack::DSL.Normalizer(),
@@ -119,16 +114,13 @@ module Trailblazer
           terminus: Linear::Normalizer::Terminus.Normalizer(),
         )
 
-        def self.OptionsForSequenceBuilder(**options)
+        def options_for_sequence_build(fail_fast_end: Activity::End.new(semantic: :fail_fast), pass_fast_end: Activity::End.new(semantic: :pass_fast), **options)
+          fail_fast_terminus_options = [fail_fast_end, magnetic_to: :fail_fast, id: "End.fail_fast", normalizers: Normalizers]
+          past_fast_terminus_options = [pass_fast_end, magnetic_to: :pass_fast, id: "End.pass_fast", normalizers: Normalizers]
 
-          options = Railway::DSL.OptionsForSequenceBuilder(**options)
+          railway_options, railway_termini = Railway::DSL.options_for_sequence_build(**options)
 
-          initial_sequence = DSL.initial_sequence(**options)
-
-          {
-            **options,
-            sequence: initial_sequence,
-          }
+          return railway_options, railway_termini + [fail_fast_terminus_options, past_fast_terminus_options]
         end
       end # DSL